refactor: consolidate workflow tags into a single component
