+++ apps/web/src/samples/components/WorkflowTags.tsx
@@ -1,13 +1,10 @@
 import { cn } from "@app/cn";
+import { getWorkflowDisplayName } from "@app/utils";
 import Icon from "@base/Icon";
+import Link from "@base/Link";
 import type { WorkflowFilterState } from "@samples/utils";
-import type { WorkflowState } from "@virtool/contracts";
-import { workflowStateIcons } from "../Filter/workflowStateIcons";
-
-type SampleItemWorkflowTagProps = {
-	displayName: string;
-	workflowState: WorkflowState;
-};
+import type { SampleWorkflows, WorkflowState } from "@virtool/contracts";
+import { workflowStateIcons } from "./Filter/workflowStateIcons";
 
 const filterStates: Record<WorkflowState, WorkflowFilterState> = {
 	complete: "ready",
@@ -22,6 +19,11 @@ const tagClassNames: Record<WorkflowFilterState, string> = {
 	none: "bg-gray-100 text-gray-600 group-hover:bg-gray-200",
 };
 
+type WorkflowTagProps = {
+	displayName: string;
+	workflowState: WorkflowState;
+};
+
 /**
  * An inline segment displaying the current state of a single workflow.
  *
@@ -29,10 +31,7 @@ const tagClassNames: Record<WorkflowFilterState, string> = {
  * @param workflowState - current state of the workflow
  * @returns A tag displaying the state of a workflow
  */
-export default function WorkflowTag({
-	displayName,
-	workflowState,
-}: SampleItemWorkflowTagProps) {
+function WorkflowTag({ displayName, workflowState }: WorkflowTagProps) {
 	const state = filterStates[workflowState];
 	const { className, icon } = workflowStateIcons[state];
 
@@ -48,5 +47,38 @@ export default function WorkflowTag({
 			<Icon icon={icon} className={cn("size-4", className)} />
 			<span>{displayName}</span>
 		</div>
+	);
+}
+
+type WorkflowTagsProps = {
+	id: number;
+	workflows: SampleWorkflows;
+};
+
+/**
+ * Workflow tags for a sample item
+ *
+ * The tags show the state of every analysis workflow associated with the sample.
+ *
+ *
+ * @param id - the sample's id
+ * @param workflows - the workflows object for the sample
+ * @returns The workflow tags for a sample.
+ */
+export default function WorkflowTags({ id, workflows }: WorkflowTagsProps) {
+	return (
+		<Link
+			className="group flex items-stretch"
+			to="/samples/$sampleId/analyses"
+			params={{ sampleId: String(id) }}
+		>
+			{Object.entries(workflows).map(([key, value]) => (
+				<WorkflowTag
+					key={key}
+					displayName={getWorkflowDisplayName(key)}
+					workflowState={value}
+				/>
+			))}
+		</Link>
 	);
 }
